KEA: add support for /vsi file systems
