+++ gdal/frmts/kea/frmt_kea.html
@@ -51,6 +51,10 @@ of the underlying HDF5 format.</p>
 <li><p><b>THEMATIC</b>=YES/NO: If YES then all bands are set to thematic. Defaults to NO</li>
 </ul>
 
+<h2>VSI Virtual File System API support</h2>
+
+Since GDAL 2.5.0, read-only operations on /vsi file systems are supported.
+
 <h2>See Also:</h2>
 
 <ul>
